Extract RType operations from Types.Types
